Why am I revisiting this 3 years later?
It was poorly thought out way back when (last pushed in 2011!), and I had a moment
of shame when looking at the state of this public repo.
As I am bored, or on the phone, or instead of doing innumerable other wastes of time
perhaps I'll continue working on this, and do something about the stupid and
terrible tkinter stuff I tacked on to this for some unknown reason.